Version 2.0.0
FossilOrigin-Name: 4e4a7fc46fd720ff81106ff679232444b43ad840063f9cbcbd01d14541ad2c60

+++ lib/vallist.cpp
@@ -35,7 +35,7 @@ using std::string;
 namespace mysqlpp {
 
 void
-create_vector(int size, std::vector<bool>& v, bool t0, bool t1, bool t2,
+create_vector(size_t size, std::vector<bool>& v, bool t0, bool t1, bool t2,
 		bool t3, bool t4, bool t5, bool t6, bool t7, bool t8, bool t9,
 		bool ta, bool tb, bool tc)
 {
